requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
httpx[socks]==0.26.0
pydantic==2.5.3
python-multipart==0.0.6
apscheduler==3.10.4
jinja2==3.1.3
greenlet==3.3.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic-settings==2.1.0

